Adding "Privacy" page.
This update introduces a new "Privacy" page that explains how the user's data are saved and used in the web app.
Now the "Privacy" page will be opened in the existing tab and no more in a new one.
The CSS stylesheet has been updated, too, introducing a new container dedicated to the "Privacy" page.