Fixes to HTML & JS.
This update adds a small popup when hovering the Dark Mode button to tell the user what it does.
Now, if the user double clicks the Dark Mode button, it resets the JSON file that stores the user's preference, which causes the page to automatically set the Mode.
Comments have been added to the JS file for better accessibility.