Introducing a "Needs to Save" reminder, a new "Help" page, & more.
This update adds a "Needs to Save" reminder, which causes the "Save your list" button to turn red when the user makes a change on their list.
In addition, a new "Help" page has been updated to help new users to understand how the web app works.
The "Privacy" page has been updated to improve the accessibility on mobile.
The "README.md" file has been updated to explain to other developers what the web app does and how they can contribute to the project in a better way.
This update includes minor bug fixes and changes to the "darkMode.js" file and others, too.