Fixes to HTML & CSS.
This update fixes a bug in the homepage that showed an empty item when the user reloaded the page after saving the list.
This update includes some improvements to the Dark Mode, too: the colors of the buttons have been changed to make them more readable on low light conditions.
[Version 5.1]