Introducing Italian language support to My Tasks.
This update introduces the support for the Italian language.
From now on, users will be able to select the language they would like to be shown in the app in the footer, using a selector.
If the user changes language, when they will open the app again in the future they will be redirected in the chosen version of it.

[Version 6.0]